Changes cluster import to read multiple params.
When running find all clsuters you can set multiple parameters and add
these to the metadata table. I have started using this feature so now I
have cluster calls for all of these different settings. This changes the
import to bring in a DataFrame of clusters instead of a series.