Apoptosis (#100)
* Looking for apoptosis datasets.

We know that germline have a slightly different apoptosis machinery. We
are curious to see if some of our clusters are enriched in dying cells,
but don't know exactly what those signals would look like. Here I am
looking for a public RNA-Seq dataset that may have induced an apoptosis
response in spermatagonia or spermatocytes. I only found one possible
dataset that KD mip-40 which is a supressort of lin-52 which has some
role in apoptosis. May be worth checking out.

* notebook to look for apoptotic GO terms.