Changes cluster colors to use shades for development.
Everyone was having issues with the color scheme for the clusters. With
20 groups it becomes difficult to separate the clusters. Instead it was
suggested to change the lineage clusters into shades of a single color.
This commit makes the germline clusters shades of red and the cyst cell
clusters shades of blue. I also tweaked the pigment cell and terminal
epithelium cells to be different enough from red and blue.