Fixes some path issues with original files.
Cell ranger creates output in the current working directory. To make the output in the correct location I move to where I want the output before running the command. This makes relative paths off by two levels `../../`. Originally I was adding `../../../` to the original filename in the sample table, but this was confusing. This changes it to prefixing the `../../{fname}` in the cell ranger command.